test: tweak serve tests to hopefully be more reliable

+++ cmd/serve_test.go
@@ -18,9 +18,12 @@ package cmd
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"net"
 	"net/http"
 	"os"
+	"strconv"
 	"strings"
 	"syscall"
 	"testing"
@@ -49,7 +52,7 @@ func init() {
 	}
 }
 
-func coreServeTest(t *testing.T, cfg string, url string) (*http.Response, error, func()) {
+func coreServeTest(cfg string, port int, url string) (*http.Response, func(), error) {
 	exitStatus = -1
 	rootCmd.ResetFlags()
 	seedCmd.ResetFlags()
@@ -63,25 +66,52 @@ func coreServeTest(t *testing.T, cfg string, url string) (*http.Response, error,
 
 	go func() { bindErr = rootCmd.Execute() }()
 
-	time.Sleep(2 * time.Second)
+	if bindErr != nil {
+		return nil, nil, bindErr
+	}
+
+	time.Sleep(time.Second)
 
 	if bindErr != nil {
-		return nil, bindErr, nil
+		return nil, nil, bindErr
+	}
+
+	ok := false
+	for i := 0; i < 10; i++ {
+		timeout := time.Duration(i*i*100) * time.Millisecond
+		conn, err := net.DialTimeout("tcp", net.JoinHostPort("127.0.0.1", strconv.Itoa(port)), timeout)
+		if conn != nil {
+			defer conn.Close()
+		}
+		if err == nil {
+			ok = true
+			break
+		} else {
+			fmt.Printf("Didn't connect to tcp: %v\n", err)
+		}
+	}
+
+	if bindErr != nil {
+		return nil, nil, bindErr
+	}
+
+	if !ok {
+		return nil, nil, errors.New("unable to connect to server")
 	}
 
 	req, err := http.NewRequest(http.MethodGet, url, nil)
 	if err != nil {
-		return nil, err, nil
+		return nil, nil, err
 	}
 
 	resp, err := http.DefaultClient.Do(req)
 
-	return resp, err, func() {
+	return resp, func() {
 		syscall.Kill(syscall.Getpid(), syscall.SIGUSR1)
 		if resp != nil {
 			resp.Body.Close()
 		}
-	}
+	}, err
 }
 
 func Test_ServeCommand_ExecuteInvalidPort(t *testing.T) {
@@ -95,7 +125,7 @@ layers:
       color: "FFFFFF"
 `
 
-	_, err, _ := coreServeTest(t, cfg, "http://localhost:12340/")
+	_, _, err := coreServeTest(cfg, 12340, "http://localhost:12340/")
 
 	assert.Error(t, err)
 	assert.Equal(t, 1, exitStatus)
@@ -121,7 +151,7 @@ layers:
 `
 	os.Setenv("KEY", "hunter2")
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12342/root/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12342, "http://localhost:12342/root/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -195,7 +225,7 @@ layers:
       color: "FFFFFF"
 `
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12341/")
+	resp, postFunc, err := coreServeTest(cfg, 12341, "http://localhost:12341/")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -216,7 +246,7 @@ layers:
       color: "FFFFFF"
 `
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12341/")
+	resp, postFunc, err := coreServeTest(cfg, 12341, "http://localhost:12341/")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -235,7 +265,7 @@ layers:
       name: static
       color: "FFFFFF"
 `
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12343/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12343, "http://localhost:12343/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -259,7 +289,7 @@ layers:
       color: "FFFFFF"
 `
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12344/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12344, "http://localhost:12344/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -283,7 +313,7 @@ layers:
       name: static
       color: "FFFFFF"
 `
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12345/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12345, "http://localhost:12345/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -307,7 +337,7 @@ layers:
       name: static
       color: "FFFFFF"
 `
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12346/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12346, "http://localhost:12346/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -332,7 +362,7 @@ layers:
       color: "FFFFFF"
 `
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12347/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12347, "http://localhost:12347/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -366,7 +396,7 @@ layers:
       name: static
       color: "FFFFFF"
 `
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12342/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12342, "http://localhost:12342/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -390,7 +420,7 @@ layers:
       color: "FFFFFF"
 `
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12348/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12348, "http://localhost:12348/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -411,7 +441,7 @@ layers:
       color: "FFFFFF"
 `
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12349/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12349, "http://localhost:12349/tiles/color/8/12/32")
 	if postFunc != nil {
 		defer postFunc()
 	}
@@ -439,7 +469,7 @@ layers:
       color: "FFFFFF"
 `
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12349/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12349, "http://localhost:12349/tiles/color/8/12/32")
 	defer postFunc()
 
 	assert.NoError(t, err)
@@ -491,7 +521,7 @@ layers:
 `
 	fmt.Println(cfg)
 
-	resp, err, postFunc := coreServeTest(t, cfg, "http://localhost:12341/tiles/color/8/12/32")
+	resp, postFunc, err := coreServeTest(cfg, 12341, "http://localhost:12341/tiles/color/8/12/32")
 	defer postFunc()
 
 	if err != nil {
